Improve error handling

+++ cgif_example.c
@@ -30,7 +30,7 @@ int main(void) {
   uint8_t       aPalette[] = {0xFF, 0x00, 0x00,                 // red
                               0x00, 0xFF, 0x00,                 // green
                               0x00, 0x00, 0xFF};                // blue
-  uint8_t numColors = 3;                                        // number of colors in aPalette (up to 256 possible)   
+  uint16_t numColors = 3;                                        // number of colors in aPalette (up to 256 possible)
 
   // initialize the GIF-configuration and create a new GIF
   initGIFConfig(&gConfig, "example_cgif.gif", WIDTH, HEIGHT, aPalette, numColors);
